+++ package.json
@@ -37,6 +37,7 @@
     "debug": "^4.1.0",
     "find-up": "^3.0.0",
     "fuzzy": "^0.1.3",
+    "git-remote-origin-url": "^2.0.0",
     "icafe-api": "^1.3.0",
     "inquirer": "^6.2.0",
     "inquirer-store": "^1.1.1",
