+++ src/debug.js
@@ -1,0 +1,12 @@
+/**
+ * @file debug
+ * @author Cuttle Cong
+ * @date 2018/10/24
+ *
+ */
+
+const debug = require('debug')('cz-conventional-changelog-befe')
+
+debug('debug enabled!')
+
+module.exports = debug
