--- src/types.js
@@ -1,100 +1,0 @@
-var i = require('./i18n/index')
-
-module.exports = {
-  build: {
-    title: i('types.build.title'),
-    description: i('types.build.description')
-  },
-  chore: {
-    title: i('types.chore.title'),
-    description: i('types.chore.description')
-  },
-  ci: {
-    title: i('types.ci.title'),
-    description: i('types.ci.description')
-  },
-  docs: {
-    title: i('types.docs.title'),
-    description: i('types.docs.description')
-  },
-  feat: {
-    title: i('types.feat.title'),
-    description: i('types.feat.description')
-  },
-  fix: {
-    title: i('types.fix.title'),
-    description: i('types.fix.description')
-  },
-  perf: {
-    title: i('types.perf.title'),
-    description: i('types.perf.description')
-  },
-  refactor: {
-    title: i('types.refactor.title'),
-    description: i('types.refactor.description')
-  },
-  revert: {
-    title: i('types.revert.title'),
-    description: i('types.revert.description')
-  },
-  style: {
-    title: i('types.style.title'),
-    description: i('types.style.description')
-  },
-  test: {
-    title: i('types.test.title'),
-    description: i('types.test.description')
-  },
-  temp: {
-    title: i('types.temp.title'),
-    description: i('types.temp.description')
-  },
-  Build: {
-    title: i('types.build.title'),
-    description: i('types.build.description')
-  },
-  Chore: {
-    title: i('types.chore.title'),
-    description: i('types.chore.description')
-  },
-  Ci: {
-    title: i('types.ci.title'),
-    description: i('types.ci.description')
-  },
-  Docs: {
-    title: i('types.docs.title'),
-    description: i('types.docs.description')
-  },
-  Feat: {
-    title: i('types.feat.title'),
-    description: i('types.feat.description')
-  },
-  Fix: {
-    title: i('types.fix.title'),
-    description: i('types.fix.description')
-  },
-  Perf: {
-    title: i('types.perf.title'),
-    description: i('types.perf.description')
-  },
-  Refactor: {
-    title: i('types.refactor.title'),
-    description: i('types.refactor.description')
-  },
-  Revert: {
-    title: i('types.revert.title'),
-    description: i('types.revert.description')
-  },
-  Style: {
-    title: i('types.style.title'),
-    description: i('types.style.description')
-  },
-  Test: {
-    title: i('types.test.title'),
-    description: i('types.test.description')
-  },
-  Temp: {
-    title: i('types.temp.title'),
-    description: i('types.temp.description')
-  }
-}
